refactor tsf
