+++ sktime/base/_panel/__init__.py
@@ -1,0 +1,2 @@
+# copyright: sktime developers, BSD-3-Clause License (see LICENSE file)
+"""Base classes for specific panel estimators, e.g., classifier and regressor."""
