+++ sktime/base/_panel/forest/__init__.py
@@ -1,0 +1,2 @@
+# copyright: sktime developers, BSD-3-Clause License (see LICENSE file)
+"""Base classes for time series forests."""
