--- sktime/series_as_features/base/__init__.py
@@ -1,1 +1,0 @@
-"""sktime.series_as_features.base module."""
