--- sktime/series_as_features/base/estimators/__init__.py
@@ -1,4 +1,0 @@
-"""sktime.series_as_features.base.estimators module."""
-__all__ = ["BaseTimeSeriesForest"]
-
-from ._ensemble import BaseTimeSeriesForest
