--- sktime/series_as_features/base/estimators/interval_based/__init__.py
@@ -1,7 +1,0 @@
-"""Interval based series-as-features estimators."""
-
-__all__ = ["BaseTimeSeriesForest"]
-
-from sktime.series_as_features.base.estimators.interval_based._tsf import (
-    BaseTimeSeriesForest,
-)
