--- sktime/series_as_features/base/estimators/tests/__init__.py
@@ -1,1 +1,0 @@
-"""Tests for sktime.series_as_features.base.estimators module."""
